Miscellaneous Tweaks
- Added timer cynchronization code checks for red alert and intercom
flashers
-- A_CheckAlert function now returns 'tics until nearest 6-second
window' so that all actor-based alert sounds and flashes start on the
same tic
-- The original change reduces lag significantly in levels with red
alert and intercom actors; this eliminates the mismatched cacophany of
alert sounds that resulted, since actors aren't spawned identically,
even with +SYNCRONIZED set
- Changed usable decorations to dynamically un-set +NOTAUTOAIMED if they
have a special set, so they show up with brackets in on-screen scans
-- Added some level-specific tag names in LANGUAGE for some button
actors
- Tweaked LaserBeam actor to respect initial spawn z-height when warping
to the master actor's location.  This allows the same laser actors to be
used for the player and for monster actors
- Added handling for 'swimming' for base Trek enemy and Klingons.  If
they're in ater, they get +FLOAT and +NOGRAVITY set, and removed once
they're out of the water
- Added some waterlevel checks to various projectiles so they die in
water
- Setting 'g_debug' cvar to 1 now gives some debug data on scannable
objects.  Not particularly useful unless you're making maps and want the
language string name generated for you.  If only we could set custom
thing strings in the editor and access them via ACS in-game...